// src/app/modules/content/content.enum.ts
export enum progressAlertFrequeincyEnum {
     weekly = 'weekly',
     biweekly = 'biweekly',
     monthly = 'monthly',
}

export enum progressAlertDayEnum {
     monday = 'monday',
     tuesday = 'tuesday',
     wednesday = 'wednesday',
     thursday = 'thursday',
     friday = 'friday',
     saturday = 'saturday',
     sunday = 'sunday',
}